Extract validation info utilities into common
Extract validation info utilities from wizardTransition module
into validationInfoUtils in common module
