+++ src/common/types/clusters.ts
@@ -1,5 +1,5 @@
 import { IRow } from '@patternfly/react-table';
-import { Cluster, ClusterUpdateParams, ClusterValidationId, Host } from '../api/types';
+import { Cluster, ClusterUpdateParams, ClusterValidationId } from '../api/types';
 import { Validation as HostValidation } from './hosts';
 
 export type Validation = Omit<HostValidation, 'id'> & {
