+++ src/ocm/components/hosts/HardwareStatus.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
-import { Host, HostStatus } from '../../../common';
-import { ValidationsInfo } from '../../../common/types/hosts';
 import {
   getFailingClusterWizardSoftValidationIds,
   getWizardStepHostStatus,
   getWizardStepHostValidationsInfo,
-  wizardStepsValidationsMap,
-} from '../clusterWizard/wizardTransition';
+  Host,
+  HostStatus,
+} from '../../../common';
+import { ValidationsInfo } from '../../../common/types/hosts';
+import { wizardStepsValidationsMap } from '../clusterWizard/wizardTransition';
 import { AdditionalNTPSourcesDialogToggle } from './AdditionaNTPSourceDialogToggle';
 
 type HardwareStatusProps = {
@@ -21,9 +22,16 @@ const HardwareStatus: React.FC<HardwareStatusProps> = (props) => {
     wizardStepsValidationsMap,
     props.host,
   );
-  const validationsInfo = getWizardStepHostValidationsInfo(props.validationsInfo, 'host-discovery');
-  const sublabel = getFailingClusterWizardSoftValidationIds(validationsInfo, 'host-discovery')
-    .length
+  const validationsInfo = getWizardStepHostValidationsInfo(
+    props.validationsInfo,
+    'host-discovery',
+    wizardStepsValidationsMap,
+  );
+  const sublabel = getFailingClusterWizardSoftValidationIds(
+    validationsInfo,
+    'host-discovery',
+    wizardStepsValidationsMap,
+  ).length
     ? 'Some validations failed'
     : undefined;
