+++ src/ocm/components/hosts/NetworkingStatus.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import { HostNetworkingStatusComponentProps, HostStatus } from '../../../common';
 import {
   getFailingClusterWizardSoftValidationIds,
   getWizardStepHostStatus,
   getWizardStepHostValidationsInfo,
-  wizardStepsValidationsMap,
-} from '../clusterWizard/wizardTransition';
+  HostNetworkingStatusComponentProps,
+  HostStatus,
+} from '../../../common';
+import { wizardStepsValidationsMap } from '../clusterWizard/wizardTransition';
 import { AdditionalNTPSourcesDialogToggle } from './AdditionaNTPSourceDialogToggle';
 
 const NetworkingStatus: React.FC<HostNetworkingStatusComponentProps> = (props) => {
@@ -14,8 +15,16 @@ const NetworkingStatus: React.FC<HostNetworkingStatusComponentProps> = (props) =
     wizardStepsValidationsMap,
     props.host,
   );
-  const validationsInfo = getWizardStepHostValidationsInfo(props.validationsInfo, 'networking');
-  const sublabel = getFailingClusterWizardSoftValidationIds(validationsInfo, 'networking').length
+  const validationsInfo = getWizardStepHostValidationsInfo(
+    props.validationsInfo,
+    'networking',
+    wizardStepsValidationsMap,
+  );
+  const sublabel = getFailingClusterWizardSoftValidationIds(
+    validationsInfo,
+    'networking',
+    wizardStepsValidationsMap,
+  ).length
     ? 'Some validations failed'
     : undefined;
